// src/app/Model/model.Loan.ts
export class Loan {
	loanId: number;
	loanAmount: number;
	rateOfInterest: number;
	tenure: number;
	marginAmount: number;
	emiAmount: number;
	processingFees: number;
	otherCharges: number;
	totalAmount: number;
	leadNo: string;
	loanNo: string;
	loanType: string;
	// tslint:disable-next-line: indent
	status: number;	// CustomerDetails customerDetails;
}